% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bowtie2Align.R
\name{Bowtie2Align}
\alias{Bowtie2Align}
\title{Bowtie2Align()}
\usage{
Bowtie2Align(
  Reference_Genome = "No Default",
  Forward_Fastq = "No Default",
  Reverse_Fastq = "No Default",
  Paired_End_Data = TRUE,
  Index_File_Name = "index",
  Output_Base_Name = "Bowtie2_Alignment",
  Threads = 4,
  flag = "--rg-id 1 --rg SM:Sample --rg LB:1 --rg PI:400 --rg PL:ILLUMINA"
)
}
\arguments{
\item{Reference_Genome}{The reference genome unto which the sequence data will be mapped (including .fasta or .fna file extension).}

\item{Forward_Fastq}{The single-read or forward paired-end fastq file name (including .fastq file extension)}

\item{Reverse_Fastq}{When Paired_End_Data=TRUE, the reverse paired-end fastq file name (including .fastq file extension). This argument is ignored when Paired_End_Data=FALSE.}

\item{Paired_End_Data}{The input data type, where FALSE corresponds with single-end reads. \link{default=TRUE}}

\item{Index_File_Name}{The base name for the Bowtie2 reference index. If no index was made before, a new one will be constructed. Otherwise, the index with this basename will be used.}

\item{Output_Base_Name}{The base name for the alignment files and reports. No file extension is required. \link{default=Bowtie2_Alignment}}

\item{Threads}{The number of threads to use in alignment calculations. \link{default=2}}

\item{flag}{Alignment data is supplied with data information such as read group identifiers (--rg-id), Sample name (--rg SM:), library (rg LB:), Predicted median insert size (--rg PI:) and platform (--rg PL:). \link{default="--rg-id 1 --rg SM:Sample --rg LB:1 --rg PI:400 --rg PL:ILLUMINA"}}
}
\description{
Based on Rbowtie2 and RSamtools, this function is used to map both single and paired-end sequence data to a reference genome, followed by converting the SAM output into a sorted and indexed BAM file.
}
\examples{
Basic Alignment of paired-end example data on a local machine (Low number of threads):
Bowtie2Align(Reference_Genome = "example.fasta",Forward_Fastq = "example_fw.fastq",Reverse_Fastq = "example_rv.fastq",Paired_End_Data = T, Index_File_Name = "example", Output_Base_Name = "example", Threads=1)
Basic Alignment of single-end example data on a server (High number of threads):
Bowtie2Align(Reference_Genome = "example.fasta",Forward_Fastq = "example_fw.fastq",Paired_End_Data = F, Index_File_Name = "example", Output_Base_Name = "example", Threads=24)
}
\keyword{alignment}
\keyword{bowtie2}
\keyword{fastq}
