% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATKCall.R
\name{GATKCall}
\alias{GATKCall}
\title{GATKCall()}
\usage{
GATKCall(
  Reference_Genome = "No Default",
  Alignment_File_Name = "No Default",
  Output_Base_Name = "Gatk_Call",
  gvcf.mode = FALSE,
  BAM_Validation = FALSE,
  BAM_Repair = FALSE,
  pathtobash = "C:/Program Files/Git/bin/bash.exe",
  pathtoGATK = "./gatk",
  pathtoSamtools = "./samtools"
)
}
\arguments{
\item{Reference_Genome}{The reference genome unto which the sequence data has been mapped (including .fasta or .fna file extension).}

\item{Alignment_File_Name}{The file name of the alignment data from which variants are to be called (including .bam )}

\item{Output_Base_Name}{The desired output base name for the variant call format (VCF) file. \link{default="Gatk_Call"}}

\item{gvcf.mode}{Whether to store the variant call data in a genomic file (g.vcf). Use this when calling variants from multiple sample alignments that are to be merged later into one large vcf. \link{default=FALSE}}

\item{BAM_Validation}{Whether to check for anomalies in the alignment file before performing variant calling. \link{default=FALSE}}

\item{BAM_Repair}{Whether to repair or replace the alignment file header. Enable this if the .BAM file is malformed or if the sample name is incorrect by inserting the new header as a character. An example of such a header is "-SM Example -LB 1 -PL Illumina -PU 10578252" \link{default=FALSE}}

\item{pathtobash}{Only for running GATK within a docker (Mainly for windows users): the file path to git bash (bash.exe) when installed in a custom file location. Linux users can alter this directory path to make use of GATK within the docker environment as well.}

\item{pathtoGATK}{Only for running GATK on linux: the file path to the GATK software when a custom installation path has been used. The default is ./gatk, which assumes gatk is added to your $PATH.}

\item{pathtoSamtools}{Only for running GATK on linux when BamValidation=T or BamRepair=T: the file path to the samtools software when a custom installation path has been used. The default is ./samtools, which assumes samtools is added to your $PATH.}
}
\description{
This function allows the user to send commands to the GATK, running either natively on linux and Mac or within a docker for Windows, from within R.
}
\examples{
# Simple call
GATKCall(Reference_Genome="ref.fasta",Alignment_File_Name="my_alignment.bam",Output_Base_Name="results")
}
\keyword{GATK}
\keyword{call}
\keyword{variant}
