% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVcf.R
\name{ReadVcf}
\alias{ReadVcf}
\title{ReadVcf()}
\usage{
ReadVcf(
  vcf_file = "No Default",
  contig = "No Default",
  nrows = -1,
  skip = 0,
  sample_nr = 0,
  verbose = TRUE
)
}
\arguments{
\item{vcf_file}{The file to be read in (including .vcf file extension)}

\item{contig}{The name of the contig that has to be read in. For example "C1" or "NC_044370.1". Note that nothing is read in if the contig name is misspelled and only one contig name can be submitted.}

\item{nrows}{The number of rows. The Scan_Vcf function returns the nrows per chromosome.}

\item{skip}{The number of rows to skip before starting to read. The Scan_Vcf function returns the skip number corresponding with the chromosome of interest.}

\item{sample_nr}{The number(s) of the sample(s) to be included. The first sample always has sample_nr 10 and the default sample_nr 0 reads in all samples present. Specific sample numbers are returned with the Scan_Vcf function}
}
\description{
Based on vcfR, this function is used to read variant data from a variant call format (VCF) file into a comprehensive list of dataframes. This list is split into general parameters and sample specific parameters.
}
\examples{
# Read only the first 10 rows of the second sample into an R object:
data<-ReadVcf("Vcf_file.vcf",nrows=10,skip=0,sample_nr=11)
# Read only the second 50 bp chromosome for all samples into an R object:
data<-ReadVcf("Vcf_file.vcf",nrows=50,skip=50,sample_nr=0)
# Read in an entire file (not recommended due to memory limitations):
data<-ReadVcf("Vcf_file.vcf")
}
\keyword{vcf}
\keyword{vcfR}
