% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantFilter.R
\name{VariantFilter}
\alias{VariantFilter}
\title{VariantFilter()}
\usage{
VariantFilter(
  vcf_list = "No Default",
  General_filter = "No Default",
  Sample_Specific_Filter = "No Default",
  min_dist = "No Default",
  min_AD = 0,
  max_AD = Inf
)
}
\arguments{
\item{vcf_list}{R object into which vcf data was stored with ReadVcf()}

\item{General_filter}{The filter criteria for general parameters, defined as a character string per criterium. View examples for details}

\item{Sample_Specific_Filter}{The filter criteria for sample parameters, defined as a character string per criterium. View examples for details}

\item{min_dist}{The minimum variant free space around any given variant in basepairs. Use this when the presence of nearby variants is undesirable, for example during variant genotyping primer construction. This filtering criterium takes all variants of the vcf into account. If only high quality variants should be taken into account, perform the filtering step twice.}

\item{min_AD}{The minimum summed allelic depth allowed. This is a measure of coverage used to discard variant with little evidende.}

\item{max_AD}{The maximum summed allelic depth allowed. This is a measure of oversequencing often observed for repeats and overabundant transposons.}
}
\description{
Based on vcfR, this function is used to read variant data from a variant call format (VCF) file into a comprehensive list of dataframes. This list is split into general parameters and sample specific parameters.
}
\examples{
# Select only SNPs (note the double == sign)
VariantFilter(vcf_list=vcf_object,General_Filter="Type=='SNP'"))
# Select homozygotes with MQ higher than 2
VariantFilter(vcf_list=vcf_object, General_Filter="MQ>2",Sample_Specific_Filter="GT=='1/1'")
}
\keyword{filter}
\keyword{vcf}
\keyword{vcfR}
