% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantToolsCall.R
\name{VariantToolsCall}
\alias{VariantToolsCall}
\title{VariantToolsCall()}
\usage{
VariantToolsCall(
  Reference_Genome = "No Default",
  Alignment_File_Name = "No Default",
  Output_Base_Name = "VariantToolsCall.vcf",
  minBaseQuality = 10,
  minMapQuality = 10,
  minDepth = 2,
  maxDepth = Inf,
  p_lower = 0.2,
  p_error = 0.001,
  read_count = 2
)
}
\arguments{
\item{Reference_Genome}{The reference genome unto which the sequence data has been mapped (including .fasta or .fna file extension).}

\item{Alignment_File_Name}{The file name of the alignment data from which variants are to be called (including .bam )}

\item{Output_Base_Name}{The desired output base name for the variant call format (VCF) file. \link{default="VariantTools_Call"}}

\item{minBaseQuality}{Minimum nucleotide quality below which a variant will be masked. \link{default=10}}

\item{minMapQuality}{Minimum mapping quality below which a variant will be masked. \link{default=10}}

\item{minDepth}{Minimum read depth below which a variant will be masked \link{default=2}}

\item{maxDepth}{Maximum read depth above which a variant will be masked \link{default=0.2}}

\item{p_lower}{From VariantTools: 'The lower bound on the binomial probability for a true variant.'}

\item{p_error}{From VariantTools: 'The binomial probability for a sequencing error (default is reasonable for Illumina data with the default quality cutoff).' \link{default=0.001}}

\item{read_count}{From VariantTools: 'Require at least this many high quality reads with the alternate base. The default value is designed to catch sequencing errors where coverage is too low to rely on the LRT. Increasing this value has a significant negative impact on power.' \link{default=2}}
}
\description{
This function allows the user to send commands to the GATK, running either natively on linux and Mac or within a docker for Windows, from within R.
}
\examples{
# Perform a simple call
VariantToolsCall(Reference_Genome="Ref.fasta",Alignment_File_Name="my_alignment.bam",Output_Base_Name="Variants")
}
\keyword{VariantTools}
\keyword{call}
\keyword{variant}
